Add new drawing object type - text
This is the same as the text_box type, but is simpler to use, in that
you don't have to specify a width and height, as there's no box.

Internal implementation details also mean now that you can pass the full
set of kwargs for the text type as supported by matplotlib.  This is a
significant benefit.